python/frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["Douglas Soo <dsoo@alum.mit.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
werkzeug = "^1.0.1"
honeycomb-beeline = "^2.12.2"
[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
